support vue-router in plugin test app
